Download historic Bioclimatic indicators (BIOs) named CMCC-BioClimInd.

Parse historic CMCC-BioClimInd bioclimatic indicators optionally with a setting of boundary and a few other options.

Usage

cmcc_bioclim(bry = NULL, path = NULL, nm_mark = "clip", return_stack = TRUE)

Arguments

bry

(sf) The boundary to mask the downloaded original data. If NULL, it would get global map. If not NULL, it can take sf, sfc, etc. The default is NULL.

path

(character) The path to save the downloaded imagery. If NULL, it would use the current working directory. The default is NULL.

nm_mark

(character) the name mark of clipped images. The default is "clip". It would be ignored if bry is NULL.

return_stack

(logical) if TRUE, stack the imagery together and return. If the area is large and resolution is high, it is better not to stack them. The default is TRUE.

Value

if return_stack is TRUE, the images would be returned as a stars. Otherwise, nothing to return, but the user would receive a message of where the images are.

Details

Web page page for this dataset

Note

The function is experimental at the moment, because the download server of this dataset is not as stable as Worldclim yet. If it fails due to slow internet, try to set a larger timeout option, e.g., using options(timeout = 1e3).

References

Noce, Sergio, Luca Caporaso, and Monia Santini."A new global dataset of bioclimatic indicators. "Scientific data 7.1 (2020): 1-12. doi:10.1038/s41597-020-00726-5

Examples

if (FALSE) { # \dontrun{
library(dplyr)
library(sf)
library(itsdm)
bry <- st_polygon(
  list(rbind(c(29.34, -11.72), c(29.34, -0.95),
             c(40.31, -0.95), c(40.31, -11.72),
             c(29.34, -11.72)))) %>%
  st_sfc(crs = 4326)

cmcc_bios <- cmcc_bioclim(bry = bry,
  nm_mark = 'tza', path = tempdir())
} # }