Download future Bioclimatic indicators (BIOs) named CMCC-BioClimInd.

Parse future CMCC-BioClimInd bioclimatic indicators obtained by different Earth System Models (ESMs) optionally with a setting of boundary and a few other options.

Usage

future_cmcc_bioclim(
  bry = NULL,
  path = NULL,
  esm = "CMCC-CESM",
  rcp = 85,
  interval = "2040-2079",
  nm_mark = "clip",
  return_stack = TRUE
)

Arguments

bry

(sf) The boundary to mask the downloaded original data. If NULL, it would get global map. If not NULL, it can take sf, sfc, etc. The default is NULL.

path

(character) The path to save the downloaded imagery. If NULL, it would use the current working directory. The default is NULL.

esm

(character) The option for Earth System Models (ESMs). Should be one of "CMCC-CESM", 'GFDL-ESM2M', 'HadGEM2-ES', 'IPSL-CM5A-LR', 'MIROC-ESM-CHEM', 'NorESM1-M'. The default is CMCC-CESM.

rcp

(numeric) The option of Representative Concentration Pathways (RCPs). Should be 45 or 85. Only 85 is available for CMCC-CESM. The default is 85.

interval

(character) The option for time interval. Should be one of "2040-2079", "2060-2099". The default is "2040-2079".

nm_mark

(character) the name mark of clipped images. The default is "clip". It would be ignored if bry is NULL.

return_stack

(logical) if TRUE, stack the imagery together and return. If the area is large and resolution is high, it is better not to stack them. The default is TRUE.

Value

if return_stack is TRUE, the images would be returned as a stars. Otherwise, nothing to return, but the user would receive a message of where the images are.

Details

https://doi.pangaea.de/10.1594/PANGAEA.904278?format=html#download

Note

The function is experimental at the moment, because the download server of this dataset is not as stable as Worldclim yet. If it fails due to slow internet, try to set a larger timeout option, e.g., using options(timeout = 1e3).

References

Noce, Sergio, Luca Caporaso, and Monia Santini."A new global dataset of bioclimatic indicators. "Scientific data 7.1 (2020): 1-12.doi:10.1038/s41597-020-00726-5

Examples

if (FALSE) { # \dontrun{
library(itsdm)
future_cmcc_bioclim(path = tempdir(),
  esm = 'GFDL-ESM2M', rcp = 45,
  interval = "2040-2079", return_stack = FALSE)
} # }