Print summary information from EnvironmentalOutlier object.

Display the environmental variable values comparing to the mean values of the detected environmental outliers in observations.

Usage

# S3 method for class 'EnvironmentalOutlier'
print(x, ...)

Arguments

x

(EnvironmentalOutlier) A EnvironmentalOutlier object to be messaged. It could be the return of function suspicious_env_outliers.

...

Not used.

Value

The same object that was passed as input.

See also

Examples

# \donttest{
library(dplyr)
library(sf)
library(stars)
library(itsdm)

data("occ_virtual_species")
env_vars <- system.file(
  'extdata/bioclim_tanzania_10min.tif',
  package = 'itsdm') %>% read_stars() %>%
  slice('band', c(1, 5, 12, 16))

occ_outliers <- suspicious_env_outliers(
  occ = occ_virtual_species, variables = env_vars,
  z_outlier = 5, outliers_print = 4L)

print(occ_outliers)
# }