Function to detect suspicious outliers based on environmental variables.

Run outlier.tree to detect suspicious outliers in observations.

Usage

suspicious_env_outliers(
  occ,
  occ_crs = 4326,
  variables,
  rm_outliers = FALSE,
  seed = 10L,
  ...,
  visualize = TRUE
)

Arguments

occ

(data.frame, sf, SpatialPointsDataFrame) The occurrence dataset for training. There must be column x and y for coordinates if it is a regular data.frame.

occ_crs

(numeric or crs) The EPSG number or crs object of occurrence CRS. The default value is 4326, which is the geographic coordinate system.

variables

(RasterStack or stars) The stack of environmental variables.

rm_outliers

(logical) The option to remove the suspicious outliers or not. The default is FALSE.

seed

(integer) The random seed used in the modeling. It should be an integer. The default is 10L.

...

Other arguments passed to function outlier.tree in package outliertree.

visualize

(logical) If TRUE, plot the result. The default is TRUE.

Value

(EnvironmentalOutlier) A list that contains

outliers (sf) The sf points of outliers

outlier_details (tibble) A table of outlier details returned from function outlier.tree in package outliertree

pts_occ (sf) The sf points of occurrence. If rm_outliers is TRUE, outliers are deleted from points of occurrence. If FALSE, the full observations are returned.

Details

Please check more details in R documentation of function outlier.tree in package outliertree and their GitHub.

References

Cortes, David. "Explainable outlier detection through decision tree conditioning." arXiv preprint arXiv:2001.00636 (2020).

https://github.com/david-cortes/outliertree

See also

Examples

library(dplyr)
library(sf)
library(stars)
library(itsdm)

data("occ_virtual_species")
env_vars <- system.file(
  'extdata/bioclim_tanzania_10min.tif',
  package = 'itsdm') %>% read_stars() %>%
  slice('band', c(1, 5, 12))

occ_outliers <- suspicious_env_outliers(
  occ = occ_virtual_species, variables = env_vars,
  z_outlier = 3.5, outliers_print = 4L, nthreads = 1)

occ_outliers
plot(occ_outliers)