Download environmental variables made by worldclim version 2.1.

Parse historic worldclim version 2.1 variables with a setting of boundary and a few other options.

worldclim2(
  var = "tmin",
  res = 10,
  bry = NULL,
  path = NULL,
  nm_mark = "clip",
  return_stack = TRUE
)

Arguments

var

(character) The option for the variable to download, should be one of tvag, tmin, tmax, prec, srad, wind, vapr and bio. The default is 'tmin'.

res

(numeric) The option for the resolution of image to download. Should be one of 0.5, 2.5, 5, 10 in minute degree. The default is 10.

bry

(sf or sp) The boundary to mask the downloaded original data. If NULL, it would get global map. If not NULL, it can take sf, sfc, SpatialPolygonsDataFrame, SpatialPolygons, etc. The default is NULL.

path

(character) The path to save the downloaded imagery. If NULL, it would use the current working directory. The default is NULL.

nm_mark

(character) the name mark of clipped images. The default is "clip". It would be ignored if bry is NULL.

return_stack

(logical) if TRUE, stack the imagery together and return. If the area is large and resolution is high, it is better not to stack them. The default is TRUE.

Value

if return_stack is TRUE, the images would be returned as a stars. Otherwise, nothing to return, but the user would receive a message of where the images are.

Details

Web page page for this dataset

Note

If it fails due to slow internet, try to set a larger timeout option, e.g., using options(timeout = 1e3).

References

Fick, Stephen E., and Robert J. Hijmans. "WorldClim 2: new 1-km spatial resolution climate surfaces for global land areas." International journal of climatology 37.12 (2017): 4302-4315.doi:10.1002/joc.5086

Examples

if (FALSE) {
library(sf)
library(itsdm)

bry <- sf::st_polygon(
  list(rbind(c(29.34, -11.72), c(29.34, -0.95),
             c(40.31, -0.95), c(40.31, -11.72),
             c(29.34, -11.72)))) %>%
  st_sfc(crs = 4326)

bios <- worldclim2(var = "tmin", res = 10,
  bry = bry, nm_mark = 'exp', path = tempdir())
}