Function to run extended isolation forest as SDM.

Call isolation forest and its variations to do species distribution modeling and optionally do model explanation.

isotree_po(
  occ,
  occ_test = NULL,
  occ_crs = 4326,
  variables,
  categ_vars = NULL,
  ntrees = 100L,
  sample_size = NA,
  sample_rate = 1,
  ndim = 1L,
  seed = 10L,
  ...,
  offset = 0,
  response = TRUE,
  spatial_response = TRUE,
  check_variable = TRUE,
  visualize = FALSE
)

Arguments

occ

(data.frame, sf, SpatialPointsDataFrame) The occurrence dataset for training. There must be column x and y for coordinates if it is a regular data.frame.

occ_test

(data.frame, sf, SpatialPointsDataFrame, or NULL) The occurrence dataset for independent test. The same structure as occ. If not NULL, there must be column x and y for coordinates when it is a regular data.frame. If NULL, no independent test will be used. The default is NULL.

occ_crs

(numeric or crs) The EPSG number or crs object of occurrence CRS. The default value is 4326, which is the geographic coordinate system.

variables

(RasterStack or stars) The stack of environmental variables.

categ_vars

(vector of character or NULL) The names of categorical variables. Must be the same as the names in variables.

ntrees

(integer) The number of trees for the isolation forest. It must be integer, which you could use function as.integer to convert to. The default is 100L.

sample_size

(integer or NULL) Alternative argument for sample_rate. If not NULL, it should be a number for sampling size in [2, nrow(occ)]. It must be integer, which you could use function as.integer to convert to. The default is NULL. Only set either sample_size or sample_rate.

sample_rate

(numeric or NULL) Alternative argument for sample_size. If not NULL, it should be a rate for sampling size in [0, 1]. The default is NULL. Only set either sample_size or sample_rate.

ndim

(integer) ExtensionLevel for isolation forest. It must be integer, which you could use function as.integer to convert to. Also, it must be no smaller than the dimension of environmental variables. When it is 1, the model is a traditional isolation forest, otherwise the model is an extended isolation forest. The default is 1.

seed

(integer) The random seed used in the modeling. It should be an integer. The default is 10L.

...

Other arguments that isolation.forest needs.

offset

(numeric) The offset to adjust fitted suitability. The default is zero. Highly recommend to leave it as default.

response

(logical) If TRUE, generate response curves. The default is TRUE.

spatial_response

(logical) If TRUE, generate spatial response maps. The default is TRUE because it might be slow. NOTE that here SHAP-based map is not generated because it is slow. If you want it be mapped, you could call function spatial_response to make it.

check_variable

(logical) If TRUE, check the variable importance. The default is TRUE.

visualize

(logical) If TRUE, generate the essential figures related to the model. The default is FALSE.

Value

(POIsotree) A list of

model (isolation.forest) The threshold set in function inputs

variables (stars) The formatted image stack of environmental variables

pts_occ (sf) A sf of training occurrence dataset

pts_bg_occ (sf) A sf of background points for training dataset evaluation or SHAP dependence plot

pts_occ_test (sf or NULL) A sf of test occurrence dataset

pts_bg_occ_test (sf or NULL) A sf of background points for test dataset evaluation or SHAP dependence plot

var_train (sf) A sf with values of each environmental variables for training occurrence

pred_train (sf) A sf with values of prediction for training occurrence

eval_train (POEvaluation) A list of presence-only evaluation metrics based on training dataset. See details of POEvaluation in evaluate_po

var_test (sf or NULL) A sf with values of each environmental variables for test occurrence

pred_test (sf or NULL) A sf with values of prediction for test occurrence

eval_test (POEvaluation or NULL) A list of presence-only evaluation metrics based on test dataset. See details of POEvaluation in evaluate_po

prediction (stars) The predicted environmental suitability

marginal_responses (MarginalResponse or NULL) A list of marginal response values of each environmental variables. See details in marginal_response

offset (numeric) The offset value set as inputs.

independent_responses (IndependentResponse or NULL) A list of independent response values of each environmental variables. See details in independent_response

shap_dependences (ShapDependence or NULL) A list of variable dependence values of each environmental variables. See details in shap_dependence

spatial_responses (SpatialResponse or NULL) A list of spatial variable dependence values of each environmental variables. See details in shap_dependence

variable_analysis (VariableAnalysis or NULL) A list of variable importance analysis based on multiple metrics. See details in variable_analysis

Details

Please read details of algorithm isolation.forest on https://github.com/david-cortes/isotree, and the R documentation of function isolation.forest.

References

Liu, Fei Tony, Kai Ming Ting, and Zhi-Hua Zhou. "Isolation forest." 2008 eighth ieee international conference on data mining.IEEE, 2008. doi: 10.1109/ICDM.2008.17

Liu, Fei Tony, Kai Ming Ting, and Zhi-Hua Zhou. "Isolation-based anomaly detection." ACM Transactions on Knowledge Discovery from Data (TKDD) 6.1 (2012): 1-39. doi: 10.1145/2133360.2133363

Liu, Fei Tony, Kai Ming Ting, and Zhi-Hua Zhou. "On detecting clustered anomalies using SCiForest." Joint European Conference on Machine Learning and Knowledge Discovery in Databases. Springer, Berlin, Heidelberg, 2010. doi: 10.1007/978-3-642-15883-4_18

Ha riri, Sahand, Matias Carrasco Kind, and Robert J. Brunner. "Extended isolation forest." IEEE Transactions on Knowledge and Data Engineering (2019). doi: 10.1109/TKDE.2019.2947676

https://github.com/david-cortes/isotree

References of related feature such as response curves and variable importance will be listed under their own functions

See also

Examples

# \donttest{
# Using a pseudo presence-only occurrence dataset of
# virtual species provided in this package
library(dplyr)
library(sf)
library(stars)
library(itsdm)

# Prepare data
data("occ_virtual_species")
occ_virtual_species <- occ_virtual_species %>%
  mutate(id = row_number())

set.seed(11)
occ <- occ_virtual_species %>% sample_frac(0.7)
occ_test <- occ_virtual_species %>% filter(! id %in% occ$id)
occ <- occ %>% select(-id)
occ_test <- occ_test %>% select(-id)

env_vars <- system.file(
  'extdata/bioclim_tanzania_10min.tif',
  package = 'itsdm') %>% read_stars() %>%
  slice('band', c(1, 5, 12))

# Modeling
mod_virtual_species <- isotree_po(
  occ = occ, occ_test = occ_test,
  variables = env_vars, ntrees = 10,
  sample_rate = 0.6, ndim = 1L,
  seed = 123L)

# Check results
## Evaluation based on training dataset
print(mod_virtual_species$eval_train)
plot(mod_virtual_species$eval_train)

## Response curves
plot(mod_virtual_species$marginal_responses)
plot(mod_virtual_species$independent_responses,
  target_var = c('bio1', 'bio5'))
plot(mod_virtual_species$shap_dependence)

## Relationships between target var and related var
plot(mod_virtual_species$shap_dependence,
  target_var = c('bio1', 'bio5'),
  related_var = 'bio12', smooth_span = 0)

# Variable importance
mod_virtual_species$variable_analysis
plot(mod_virtual_species$variable_analysis)
# }