Isolation forest-based presence-only species distribution modeling

This package is a wrapper for a few packages including isotree, outliertree, fastshap, etc. It does purely presence-only species distribution modeling with isolation forest and variations such as SCiForest and EIF. It also provides functions to make response curves, analyze variable importance, analyze variable dependence and analyze variable contribution. As utilities, the package includes a few functions to download bioclimatic variables including worldclim version 2.0 and CMCC-BioClimInd. There are also functions to detect outliers in the occurrence dataset to do data cleaning.

Details

This package provides multiple features.

Download bioclimatic variables and reduce their dimensions. This includes historic and future climatic indicators from two sources:

worldclim

CMCC-BioClimInd

Detect suspicous environmental outliers.

Fit a isolation forest-based SDM.

Make presence-only evaluation.

Generate response curves of environmental variables including marginal and independent responses and analyze interactions between environmental variables.

Analyze variable importance using Shapley values.

Convert predicted environmental suitability to presence-absence map.

Analyze variable contributions to any specific observations.

References

Please check references in R documentation of each specific function.

Author

Lei Song lsong@clarku.edu

Maintainer: Lei Song lei.song@rutgers.edu